requirements.txt
Flask==3.0.0
supabase==2.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
pandas==2.2.0
openpyxl==3.1.2
